refactor: moved models and schemas to top of module than nesting it inside state
