+++ backend/core/comic_builder/api/dependencies.py
@@ -8,7 +8,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from core.auth.dependencies import get_current_user_id
 from core.services.database import get_async_db_session
 
-from ..state.relational.models import Project
+from ..models import Project
 
 
 async def verify_project_access(
