+++ backend/core/comic_builder/api/projects.py
@@ -9,8 +9,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from core.auth.dependencies import get_current_user_id
 from core.services.database import get_async_db_session
 
-from ..state.relational.models import Project
-from ..state.relational.schemas import (
+from ..models import Project
+from ..schemas import (
     ProjectCreateSchema,
     ProjectListResponseSchema,
     ProjectResponseSchema,
