+++ backend/core/comic_builder/models/character.py
