+++ backend/core/comic_builder/models/panel.py
