+++ backend/core/comic_builder/schemas/character.py
