+++ backend/core/comic_builder/schemas/panel.py
