+++ backend/core/comic_builder/schemas/story.py
